backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
Pillow==11.1.0
PyPDF2==3.0.1
pandas==2.2.3
pdfplumber==0.11.5
openpyxl==3.1.5
